+++ changelog.md
@@ -1,11 +1,11 @@
-# Version 2024.8.15 (2024-08-26)
+# Version 2024.8.15 (2024-08-29)
 
 - Avoid permanent cache save on remove device
 - Check rx_mode
 - Ensure only one load/save of cache file at time
 - Small definition fix for DALI
-- Use ParameterData for paramset descriptions
-- Use Typed dict for device_description
+- Use TypedDict for device_description
+- Use TypedDict for parameter_data
 
 # Version 2024.8.14 (2024-08-26)